More AMD64 fixes (release build works now, and fixed verbosity bug).

+++ js/src/nanojit/Fragmento.h
@@ -238,7 +238,7 @@ namespace nanojit
 #else
     inline int nbr(LInsp x)
     {
-        return int(x) & (NJ_PAGE_SIZE-1);
+        return (int)(intptr_t(x) & intptr_t(NJ_PAGE_SIZE-1));
     }
 #endif
 }
